+++ beacon-formats/src/netcdf/source/schema_cache.rs
